backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jenish Bhuju <jnishvuzu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.2"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
